build: Move device specific recovery files to vendor_boot
When building with GKI and vendor_boot, boot.img should be generic,
however as recovery-as-boot is mandated, the recovery ramdisk on
boot results in it not being generic. This makes use of the
initramfs on vendor_boot.img to hold all device specific recovery
assets.

This change also moves vendor, odm and product properties out of
default.prop in order to keep that generic as well. They're moved
to vendor.prop, with a corresponding change in init to read that
file if present in recovery.

Test: lunch aosp_quill-eng; m -j bootimage vendorbootimage
      Flash, validate recovery boots
Change-Id: Idda8edd7976c6428cf6253d3b6ec13e4f4aac408