Parse the runtime fingerprint prefixes from build prop
In practice, some partners use the 'import' statement to override
the device fingerprint at runtime. This CL supports parsing a subset
of import statement that the init process recognizes. And we loose
the restriction based on how the dynamic fingerprint is used in
practice.

Right now, we only searches for the override of ro.product.name and
ro.product.device. Also the current supported 'import' statement should
only include one placeholder. And the placeholder format should be
${placeholder}, with its value supplied by the script caller.

Bug: 152167826
Test: unittests pass
Change-Id: Iac093a40dc6f873c5e1858efa44cb2bd6082508a