Make required modules support :32 and :64 suffix
Support specifying bitness of required modules with :32 or :64 suffix,
like in PRODUCT_PACKAGES.

If a module is for cross host OS, the required modules are also for that
OS.
Required modules explicitly suffixed with :64 or :32 resolve to that
bitness.
Otherwise if a module is native module, the required modules resolve to
the same bitness.
Otherwise the required modules resolve to both variants, if they exist.

Bug: 155869107
Test: TH
Test: check the diff of product_target_FILES and product_host_FILES
Change-Id: Ib2f4b548535e983d621dad71cb920101111140f2