Update patch set 1
Patch Set 1:

Why is this needed in general... I am under the impression that the boot partition size means the physical block device size on the device. The size of the boot.img can be arbitrary value smaller than the physical size; and we pad it for avb signing.

Patch-set: 1
Reviewer: Gerrit User 1088085 <1088085@85c56323-6fa9-3386-8a01-6480fb634889>