Update patch set 3
Patch Set 3:

> Patch Set 3: -Code-Review
> 
> not sure if this will work yet - the boot.img won't match the boot partition size on most devices, so the avb footer won't be at the end of the boot partition

Note that in VTS compliance, we're using device-specific vbmeta.img (not aosp_*'s vbmeta.img).

So it depends on how the 'boot' desc is configured in the device-specific vbmeta.img.
It maybe a chained partition or a non-chained partition.

Right, if it's chained and GKI size != original boot.img size then the AVB footer won't be at the end of the boot partition. We can use 'avbtool resize_image' or support relocating the footer in fastboot"d" (can be a separate discussion).

Patch-set: 3
Reviewer: Gerrit User 1080813 <1080813@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0