Update patch set 19
Patch Set 19:

Btw, note that we've mandated to use "dynamic partitions" for hashtree partitions in Q. 
https://source.android.com/devices/tech/ota/dynamic_partitions/implement

And in R, we're further changing the partition layout as below.

https://docs.google.com/document/d/1J6G-uBtReYpsAWUnQrIT2HnVwRS7aq4x1a_OMop_hGQ/edit#heading=h.bv19ovzfymq7

Maybe we should think about how does custom_images fit into those categories.
e.g., it probably should be a 'dynamic partition' as well?

And if so, it cannot be a normal chained partition because the bootloader cannot recognize 'dynamic partitions' so it cannot read the AVB footer from the end of the partition.

With that, we'll have to put the 'custom' hashtree descriptor into vbmeta_system.img, which in turn means we'll have to generate multiple vbmeta_system.img, one for each custom images.

https://docs.google.com/document/d/1Dw_mClBSXr9HUKYV0594ykT74_ldWY-Op652rMfRCC8/edit#heading=h.gs4ip2uuuxrd

Or maybe we should add a vbmeta_custom.img?
(or this worth opening a bug or e-mail for initial discussion first).

Patch-set: 19
CC: Gerrit User 1061057 <1061057@85c56323-6fa9-3386-8a01-6480fb634889>
CC: Gerrit User 1290458 <1290458@85c56323-6fa9-3386-8a01-6480fb634889>
CC: Gerrit User 1108565 <1108565@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1080813 <1080813@85c56323-6fa9-3386-8a01-6480fb634889>