Update patch set 2
Patch Set 2:

When relanding a CL it's useful to revert the revert and then upload the new version on top.  That way PS1 is the previously reviewed version, and PS2 has the fixes.

Can you highlight what is different between this CL and the previous one?

Patch-set: 2
Reviewer: Gerrit User 1002751 <1002751@85c56323-6fa9-3386-8a01-6480fb634889>