Move recovery_text_res out of PACKAGING
`m installclean`, which runs in between incremental builds on our build
servers, deletes the obj/PACKAGING directory. So with recovery_text_res
in that directory, we were having to rebuild it on every incremental
build, which isn't optimal when it takes nearly 90s to build.

So with:
 $ lunch aosp_coral-eng
 $ m bootimage
 $ m installclean
 $ m bootimage

The second `m bootimage` went from executing 666 actions in 93s to
executing 658 actions in 6s.

Test: treehugger; sequence above
Change-Id: I005592e49443aab45ed039a2f0c63f7a69035565