Use checkbuild-soong to checkbuild soong modules
Have Make's checkbuild target depend on the checkbuild-soong target
instead of depending on the Make copy of Soong modules, which uses
unnecessary disk space to copy every module output file.

Also fixes checkbuild coverage of modules that are only installed
to apex files, which are not exported to Make and therefore not
visible to checkbuild.

Test: m checkbuild
Change-Id: I8fbc47995093bbcf00f638ee756575ad868a0ae6