Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I also see another place that I think can be optimized
> core/java_common.mk:389
> ALL_DEPS.$(LOCAL_MODULE).ALL_DEPS := $(ALL_DEPS.$(LOCAL_MODULE).ALL_DEPS) $(full_java_bootclasspath_libs)
> ->
> ALL_DEPS.$(LOCAL_MODULE).ALL_DEPS += $(full_java_bootclasspath_libs)

Thanks for the note!

Yeah, it looks like that's safe (but only because we've already initialized ALL_DEPS.$(LOCAL_MODULE).ALL_DEPS in base_rules.mk).

That being said, based on my measurements on AOSP, of the 328 times that we execute that line, the total execution time is 0.001s. On a larger build (~1100 times), it was at 0.003s. Switching to += drops that to 0.002s. So I'm not going to worry about that for now.


Overall, this sequence of 4 changes dropped our kati times by ~50% (and maxrss by ~30%). I don't see any obvious hotspots next, so I was going to move on to other work after getting these in (and trying to get them into Android 11). Especially as Soong is taking longer than Make now 😊

Patch-set: 1