Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> This will force any build that wants coverage to build every java library defined in the tree, not just the ones necessary for the product.

That's not strictly true. It will only build every java library that is instrumented and not every java library is instrumented. There are also existing mechanisms to filter out the modules that can be instrumented, e.g. JavaCoveragePaths and JavaCoverageExcludePaths. e.g. When I build this on AOSP the generated jacoco-report-classes-all.jar contains 71 files.

Any suggestions for how the list of jacoco jars could be filtered to further restrict it to the set of ? Or is there a better approach altogether?

Patch-set: 1
Label: Presubmit-Ready=0