Don't inherit tablet-dalvik-heap for GSI and emulator
GSI and emulator should not be specialized for tablet.

This is also to avoid the expected sysprop conflict after
I9c073a21c8257987cf2378012cadaeeeb698a4fb gets in. With the change,
duplicate assignments of a sysprop is prohibited. We currently have the
duplication due to the following hierarchy chain:

aosp_arm64.mk
 -> emulator_vendor.mk -> goldfish/vendor.mk -> phone-xhdpi-2048-dalvik-heap.mk
 -> generic_arm64/device.mk -> tablet-dalvik-heap.mk

Many of the dalvik.vm.* properties are duplicated between phone-*-
dalvik-heap.mk and tablet-dalvik-heap.mk files.

Bug: 117892318
Bug: 158735147
Test: atest --host post_process_prop_unittest

Change-Id: I4d1e2f819fe688a4a85e58387b6af58d603399d3