Adjust and dedup MODULES-IN-* logic
Notice the logic for setting up dependency from $(my_register_name)
to MODULES-IN-* is almost identical to my_path_comp.
Adjust the definition of my_path_comp and we can dedup the logic by
"rerolling" my_path_comp.

This also avoids creating cyclic dependency of `MODULES-IN-: MODULES-IN-`
when LOCAL_PATH is not set properly.

Bug: 162918277
Test: m com.android.art
Test: m MODULES-IN-art
Test: ninja -t browse
Change-Id: I7d9946a1a97e5dfeae194808aea7b6cec4335aee