Add timestamp to ramdisk.
Add a build.prop file for ramdisk. Properties uses the
name ro.[product.]bootimage*.

These ro.[product.]bootimage.* properties are also included in recovery
properties.

The file is installed to first_stage_ramdisk/boot/etc/build.prop. In
addition, a symlink boot/etc/build.prop is created to point to this
file, regardless of BOARD_USES_RECOVERY_AS_BOOT. With this setup,
aosp_arm64 can create boot images so that devices with or without
recovery_as_boot can find this file.

On devices with dedicated recovery partition, the file and symlink is
installed to ramdisk/, which is installed to the ramdisk in the boot
image. The file and symlink is NOT installed to recovery/root to prevent
collision.

On devices with recovery_as_boot, in addition to ramdisk/, it is also
installed to recovery/root, which is installed to the ramdisk in the
boot image.

Test: m bootimage and inspect output
Bug: 169169031
Bug: 162623577

Change-Id: I94b993ce3214356036d038b6db57c4e1b755c111