Rework class loader context implementation.
In Soong, the old flat representation is changed to a map of trees
(keyed on SDK versions). In Make, the new representation is not as
advanced as in Soong -- it is a map of flat lists rather than trees.
It works for now, because there is only one app that has a non-& class
loader context, and it is simple enough to use flat lists.

Test: lunch aosp_cf_x86_phone-userdebug && m
Bug: 132357300
Bug: 168686456
Change-Id: If1080295a2fafa6772383ee9fbffd7c458462c18