Rework class loader context implementation.
Changes:
 - rework internal representation of class loader contexts
 - pass serialized tree instead of list to construct_context
 - do not add shared library dependencies to the manifest_fixer
 - bring Soong class loader contexts in sync with PackageManager

The old representation consisted of a list of libraries (UsesLibraries),
a list of optional libraries (OptionalUsesLibraries) and a mapping from
library name to its build/install paths (LibraryPaths). The separation
into lists and map was necessary because of compatibility libraries that
are only added to class loader context of those apps which have
targetSdkVersion in the manifest lower than the version in which the
compatibility library was added as a separate library.

The new representation is a mapping from target SDK version to a tree
structure ClassLoaderContext. Ech node of the tree represents a library
and contains library name, build/install paths, shared flag and a slice
of subcontexts for dependencies. The same library may occur in the tree
multiple times in case it is a dependency of multiple libraries.

The new tree representation preserves class loader context structure,
which was impossible with the old flattened representation. The newly
added "shared" flag allows one to differentiate between the libraries
that need to be added to class loader context and those that need to be
added to the manifest by the manifest_fixer (the latter is a subset of
the former, not including dependencies of shared libraries).

Test: lunch aosp_cf_x86_phone-userdebug && m
Bug: 132357300
Change-Id: I49687632eb10f795cb3d92126d3b0ced7a474f1c