Add fs_config_(dirs|files) for product and system_ext
These files were added to the device with the
fs_config_files_nonsystem module in base_vendor.mk.
As they are partition-specific, move them to each base_<partition>.mk
file.

To add the fs_config_* files to the base_<partition>.mk files, it is
required to define the fs_config_* regardless of the existance of
partitions.

Bug: 170282998
Test: build and check if they are installed.
Change-Id: Ib8a2c75e2e0e93bb7030da981494e880f8465e5a