Add the connectivity APEX to product packages
The connectivity APEX contains libraries modularizing
ConnectivityService for S.

Eventually contents of the connectivity APEX will be merged into the
tethering APEX to be shipped as part of that module, however this is not
doable immediately as the tethering APEX is being shipped to production
every month, so it should not contain ConnectivityService
work-in-progress.

To solve this, use a separate connectivity APEX next to the tethering
APEX. Branches that are not used to build modules for monthly production
updates can use a consolidated APEX instead.

Bug: 171540887
Test: m, device boots and has connectivity / VPN
Change-Id: I7cf89d8002ec05cdd8e4d7f167dfe0b6194d32ab