Use GENERIC_KERNEL_CMDLINE only if GKI is used.
Explicitly guard the use of GENERIC_KERNEL_CMDLINE with
BOARD_USES_GENERIC_KERNEL_IMAGE == true. If not using GKI,
GENERIC_KERNEL_CMDLINE should not be included.

In particular, in the rare case where the device uses GKI
/ generic boot image, has a dedicated recovery
partition, and includes a kernel in the recovery partition,
also put GENERIC_KERNEL_CMDLINE in the recovery kernel cmdline.

Bug: 171512004
Test: builds

Change-Id: Ia84e604a8ded28af39c7f1861ff5d3b3af55849f