Add $OUT/boot*.tf.img
These are the boot images from the target files package. Building
these requires the whole target files package to be built in the
first place.

Expose them so that, for aosp_arm64, APEXes can be built from
boot-5.4.tf.img, not boot-5.4.img, because the latter is not
the final image being released.

Bug: 172682114
Test: m com.android.gki.kmi_5_4_android12_0_boot-5.4.tf

Change-Id: Ib97a1a9bce0de149d446871b1391c31168dbddc6