build_image can be used by soong
build_image internally calls mkuserimg_mke2fs, which again calls mke2fs.
When build_image is used in Soong, the call to mke2fs fails due to the
PATH restriction. mkuserimg_mke2fs is executed without
out/soong/host/linux-x86/bin in PATH, and therefore mke2fs is linked to
the one that is outside of Android build system. To fix the problem,
amend PATH to have the bin path.

Bug: 172414391
Test: m
Change-Id: I5bb38133a59e45e5274c4dbf7842813ac5c03d39