Move snapuserd into the vendor ramdisk.
For devices that are GKI capable, including virtual_ab_ota_compression
will now include snapuserd in the vendor ramdisk.

For devices that are not GKI-capable (eg do not have vendor_boot),
virtual_ab_ota_compression_retrofit.mk must be included in addition to
virtual_ab_ota_compression.mk. This structure is to avoid needing a
large combination of different Makefiles.

Bug: 173463595
Test: manual test
Change-Id: Idc740d3523c2d166dd578cee41ea07d7eaf8edfa