Set allow list for generic ramdisk & remove e2fsck from GKI
Devices that uses generic ramdisk must inherit from generic_ramdisk.mk.
This makefile ensures that only a set of files can be installed to the
ramdisk. Other files must be installed to the vendor-ramdisk.

Let aosp_arm64 use this makefile. As a result, e2fsck is removed from
the generic ramdisk.

Bug: 173742069
Test: manual
Change-Id: I814735156756c73a54d54c0b5e17d75cbac128b1