Exclude auto-included modules for path checks
There's a set of modules that are included based on how certain
BoardConfig variables are set -- not based on the contents of the
product makefiles.

Previously, all these auto-included modules and their deps were
completely excluded from the artifact path checks, which was hiding
problems with the path requirements in some situations.

Test: lunch aosp_arm64; m nothing (with an without path violations)
Test: with generic_ramdisk.mk in follow-up
Change-Id: Idd800768d6e22440add3a5e2736b0f5804229351