Read ro.gfx.angle.builtin.name for ANGLE's filename
The ANGLE shared object filename is currently hardcoded to
libGLESv2_angle.so, which prevents OEMs from specifying their own
filename when using ANGLE as the default OpenGL ES driver.

This CL introduces the system property ro.gfx.angle.builtin.name which
specifies the built-in ANGLE shared object filename at build time. By
default, the string will remain libGLESv2_angle.so. However, this allows
the value to be updated more easily, without forking the framework code.

The filename when loading ANGLE from an APK will remain
libGLESv2_angle.so, for compatibilty reasons. This enforces naming
conventions when loading ANGLE APKs on to the device, regardless of the
name of the built-in version of ANGLE.

Test: atest CtsAngleIntegrationHostTestCases
Change-Id: I3aee9df0b2748e3b98496deaf095a08f9184ff51