Switch platform-NDK unwinder to LLVM libunwind.a prebuilt
The new libunwind.a is distributed in Clang's resource dir, which isn't
included in the platform's prebuilts/ndk/current directory. Use the
libunwind.a directly from the toolchain prebuilt rather than through
the NDK. (We already do this for the builtins.)

For arm32, binary.mk and cxx_stl_setup.mk had been adding two different
versions of the LLVM unwinder to the same linker command-line. Remove
the duplicate one from binary.mk.

Bug: http://b/153025717
Test: device boots, manually inspect some linker command lines
Change-Id: Ie7182bd565014348b16448976dc6587d2885ba99