Configure frameworks/base/config/boot-image-profile.txt
This change is motivated by removing the frameworks/base
project from a vendor build manifest (i.e., a reduce
manifest containing only vendor projects and installed
snapshots to supply framework dependencies). In this
case, if we include boot-image-profile.txt as a prebuilt
in the java snapshot, we are able to remove a dependency
from the frameworks/base project.

Bug: 177695435
Test: Local build with alternate boot-image-profile.txt
Change-Id: I0bf41123b9b6cdf9318df22e32ffd9fced906cc3