Build prep for 2021 CTS UICC certificate change.
Log a warning at build time if the build uses the opt-out flag. This
opt-out mechanism will be removed in a later Android release, at which
point OEMs will be required to use the updated SIM cards.

For the time being, since we still haven't fully distributed the updated
SIMs internally, we opt Pixel out of this change, enaling continued use
of the old SIMs, e.g. in device labs and international offices. Later in
the S development cycle, we will remove the opt-out for Pixel.

Cuttlefish is unaffected, as the modem simulator has been updated to
support the new hash in aosp/1559881.

Bug: 178419755
Test: m CtsCarrierApiTestCases (with opt-out property set)
  (ensure deprecation warning is logged)
  keytool -printcert -jarfile $ANDROID_HOST_OUT/cts/android-cts/testcases/CtsCarrierApiTestCases.apk
Test: make a Cuttlefish build, enture no warning and CTS APK is signed
  with cts-uicc-2021-testkey
Change-Id: Id2e8eb33518cf0bde7b5a8fdcf29cbabcf426c5b