Adding a build target for devices that support RP
Launching in S, remote provisioning requires an app service that should
naturally only be included on devices that have the underlying HAL to
support remote provisioning. Adding the
remote_provisioning_system_ext.mk file will install the app in
system_ext to enable remote provisoning on device.

In the future, remote provisioning will be mandated and this file will
be deprecated, with the product package addition moving directly to
base_system.mk

Test: Devices that include this target build succesfully.
Change-Id: Id950597eb861a6c86eedd45f7caec0dbaf2d84fd