Enable a prop to determine the new figerprint is enabled
VTS and some other tests would replace the system images with GSI. To
put the correct fingerprint in the test report, we need to know if the
new fingerprint format is in use. So, add a vendor build prop.

OEMs are reponsible for setting this build prop, or using other ways
to put the correct fingerprint in the test report.

Bug: 188824341
Test: boot the device, check build prop
Change-Id: I6bc7f01903865fc2c256d209debdab68cd9d1bb3