Introduce droidcore-unbundled goal
The droidcore-unbundled goal replaces the droidcore goal
when TARGET_BUILD_UNBUNDLED_IMAGE=true. This goal is
similar to droidcore in terms of dependencies, but it
has a smaller set of dist files. The intention is to use
TARGET_BUILD_UNBUNDLED_IMAGE=true with a VSDK-using
vendor build to supply Java dependencies from prebuilts
to avoid building (or needing) Java framework sources.

Test: build with and without TARGET_BUILD_UNBUNDLED_IMAGE=true
Bug: 188176942
Bug: 187064252
Change-Id: Ia7f8200c1bb1aaaa57c177d1698a114efe26bec7