Package boot/vendorboot meta-data without the images.
Allow the meta-data (misc_inf.txt entries, ramdisks, configs) to be
packaged into a target files archive to be used later to generate the
boot and vendor_boot images.

When using partial builds, there is a merge step needed to combine
builds. It would be convinent to use the merge step with a mixed kernel
build and delay the generation of boot and vendor_boot images until
then. This requires though having all the needed meta-data from the
platform build to do so.

Bug: 191070569
Test: m dist
Change-Id: I050b2791dd7cd8ba2c8857624f000aa483c0c506