Add product variables for better control of the APEXes getting installed in the system image.
PRODUCT_INSTALL_APEXES is introduced to list the APEXes to be
installed, and it is propagated to the "installable" flags. Any APEX
not listed in PRODUCT_INSTALL_APEXES will default to installable:false,
and hence be guaranteed to not get installed into the system image.

PRODUCT_INSTALL_APEXES_OVERRIDES is also introduced to perform override
replacements on the variable above, to be able to contain the knowledge
about APEX overrides to the products that need to do them. This
somewhat duplicates the override logic through LOCAL_OVERRIDES_MODULES,
which is only applied to PRODUCT_PACKAGES in the final make install
step and hence too late to be reliable in Soong.

Test: m droid TARGET_PRODUCT=cf_x86_64_phone \
        SOONG_CONFIG_art_module_source_build={true,false}
Test: m droid TARGET_PRODUCT=aosp_cf_x86_64_phone \
        SOONG_CONFIG_art_module_source_build={true,false}
Bug: 192006406
Change-Id: I5f0290685ed707d014821fa30de034c07b3fc08c