Introduce PRODUCT_INSTALL_APEXES for APEXes to be installed.
In this CL it is no different from PRODUCT_PACKAGES, but later CLs will
move all installed APEXes to it and add consistency checks. It then
becomes the default for the "installable" flags on the apex modules in
.bp files.

PRODUCT_INSTALL_APEXES_OVERRIDES is also introduced to perform override
replacements on the variable above, to be able to contain the knowledge
about APEX overrides to the products that need to do them. This
somewhat duplicates the override logic through LOCAL_OVERRIDES_MODULES,
which is only applied to PRODUCT_PACKAGES in the final make install
step and hence too late to be reliable in Soong.

Test: m nothing
Bug: 192006406
Change-Id: I07a68082aab724a046fd0365836d369b8bf8d5ec