Read license metadata graph.
dumpgraph outputs edges of the graph as "target dependency annotations"

listshare outputs csv of projects to share to meet restricted and
reciprocal license requirements with one project per line. The first
field is the path to the project, and subsequent fields identify the
license resolutions as colon-separated target:annotations tuples.

Bug: 68860345
Bug: 151177513
Bug: 151953481

Test: m all
Test: m systemlicense
Test: m listshare
Test: m dumpgraph

Change-Id: I4ff3f98848f7e6d03a35734300d763ef5f245d53