Implement fsverity metadata signing
Using fsverity tool, artifacts in system image can be signed. Users are
to provide the following Makefile variables in device.mk to do that.

1. PRODUCT_SYSTEM_FSVERITY_SIGN_FILES
This is a boolean variable. If set to true, the following artifacts will
be signed.

- system/framework/*.jar
- system/framework/oat/<arch>/*.{oat,vdex,art}
- system/etc/boot-image.prof
- system/etc/dirty-image-objects
- system/etc/updatable-bcp-packages.txt

2. PRODUCT_SYSTEM_FSVERITY_SIGN_KEY
This is a path to a PKCS8 private key, in DER format. A test key for
development is in build/make/target/product/security/fsverity.pk8.

3. PRODUCT_SYSTEM_FSVERITY_SIGN_CERT
This is a path to a certificate, in PEM format. A test cert for
development is in build/make/target/product/security/fsverity.x509.pem.

Setting these, one fsverity metadata container file per one input file
will be generated in system.img, with a suffix ".fsv_meta". e.g.
a container file for "system/framework/foo.jar" will be
"system/framework/foo.jar.fsv_meta".

Bug: 193113311
Test: build with the following 3 Makefile variables in device.mk:
    - PRODUCT_SYSTEM_FSVERITY_SIGN_FILES
    - PRODUCT_SYSTEM_FSVERITY_SIGN_KEY
    - PRODUCT_SYSTEM_FSVERITY_SIGN_CERT
Change-Id: Ib70d591a72d23286b5debcb05fbad799dfd79b94