Add Soong variable for optimizing system Java
Introduce a Soong variable to allow selective enabling of Java
optimizations for platform targets. This will be used initially
to control optimizations (and shrinking) for system server + sysui.

Opting in to optimizations can be achieved with either:
  * Env variable (`export SYSTEM_OPTIMIZE_JAVA=true`)
  * Make variable (`SOONG_CONFIG_ANDROID_SYSTEM_OPTIMIZE_JAVA=true`)

Note that the actual framework build rule changes to hook into this
variable will land separately.

Bug: 203088572
Test: `SYSTEM_OPTIMIZE_JAVA=true m` (observe env variable change log)
Change-Id: I959985bdff394c2445cb1632116846a8d2f5443d