Initialize global variables with correct types
rbcrun can pass global variables to the starlark code from
either it's command line arguments or enviornment variables.
In both cases, it is passed as a string.

TARGET_BUILD_APPS is one such enviornment variable that gets
passed to the starlark code as a string. However, in
envsetup.rbc it is compared to an empty list, which will
return false when it really shouldn't.

Bug: 201700692
Test: Presubmits
Change-Id: I20b3694b06fdca194022310a5f30c066291fd341