Fix product-graph
product-graph was giving 'writing to readonly directory: "win_sdk"'
errors. This seems to be because win_sdk was a dist target,
but was never marked as .PHONY so kati thinks it's creating a file
called win_sdk outside of the out/ directory.

Mark win_sdk as a phony to fix the issue.

Fixes: 212724080
Test: `m product-graph` on aosp_arm64-userdebug (and other products)
Change-Id: I8b6bdf6c995a4b9ee0b09bf858b01385a591555f