Fix ota_from_target_files error in non-AB VF
For vendor freeze, vendor.img/odm.img is rebuilt and moved to
the merged target files dir in a special build flow.
https://android-review.googlesource.com/c/platform/build/+/1752968

But considering non-AB case, some more files are generated by
MakeRecoveryPatch() along with vendor.img, such as: recovery.img,
install-recovery.sh. They should be moved to the merged target files
dir as well, in order to pass the check in hasRecoveryPatch() and
GetSparseImage().

Change-Id: I003186b30e700413d54ee1b7521bd03ba83ad3f5