Make it possible to avoid specifying boot partition sizes.
Virtual platforms may have no boot partition as such, so there
is no practical limit on the partition size. Therefore, make it
possible to omit the partition size (i.e. by omitting a definition of
BOARD_*BOOTIMAGE_PARTITION_SIZE), allowing unrestricted boot partition
sizes on those platforms.

The macro get-bootimage-partition-size needs to return an empty string
in the case where the variable is not defined in order for consumers
such as assert-max-image-size and get-hash-image-max-size that test
its result using $(if) to properly disable the size check in that
case. The indentation was causing the macro to return the string
"  " instead, which caused an unrequested size check to be inserted.
Fix it by removing the indentation.

A separate patch to avbtool makes it possible to omit the
--partition_size argument.

Change-Id: Ie836feeb4c3e3b86d9c7950ff45933294cea9410
Bug: 142352330