Remove rblf.setDefault()
In make, when you $(call inherit-product), @inherit values are
appended onto all the product config variables. In Starlark,
the equivalent of the @inherit values were only added if the
product config variable had already been assigned to in this
makefile, or if rblf.setDefault() was called for it.

This creates issues with ?= assignments, where they would sometimes
do nothing in Make because of the @inherit value, but set the
value of the variable in starlark because the @inherit value
wasn't there. It also had issues with scenarios where mk2rbc
wouldn't correctly generate the rblf.setDefault() call when
it should've.

To more closely match Make and resolve these issues, append
the @inherit equivalents whenever inherit() is called, and
remove rblf.setDefault().

Bug: 222737841
Test: Manually
Change-Id: I8a4cb39adc9a5357aefd89bdff241552140c8e4a