Build orchestrator script to copy files of cc_api_surfaces
This is a helper script that uses bazel query to determine output
filepath of a cc_api_surface bazel target.
The scope of this script includes parsing the output json file and
1. copy header files to multi-tree out/
2. copy symbol files to multi-tree out/
3. generate Android.bp in multi-tree out/

Build orchestrator should be selective in running steps 1-3. ninja uses
mtime, and if the copy operation happens in every run, Soong will
always recompile the ninja graph. To overcome this, copy only if src hash
has changed by using rsync --checksum

This script has been created in build/make/tools (inner tree specific),
but will probably be moved to a different branch in the long run

Test: gen_cc_api_surface.sh $OUT_DIR //build/bazel/tests/api_surface:ndk
$ANDROID_BUILD_TOP
Bug: 213499564

Change-Id: I9fbce6b86c334a86596aeefdfe750c9926bc205d