Start script to finalize branch release.
To put in CI. At the end of every Android release, we need
to do 100,000 things to get the release ready. The process
to finalize the release for AIDL interfaces is automated now,
so we're adding a generic script. The goal is:

    1. gradually add more pieces of the finalization process
       which can be automated (VNDK, SDK, etc...)
    2. when problems happen, we can fix them in real time in CI
    3. once the entire release process is in CI, we can optimize
       it, and we can require that new requirements on the
       Android release process come with a script to satisfy them.

Until then, this script will be used to test the automated parts
of the release process in CI.

Bug: 190577319
Bug: 228979367
Test: ./build/make/finalize_branch_for_release.sh (in the T release branch)
Change-Id: Ie9ff42dd4c6e86d395bde3f14788029e5bde5fb2