dex_preopt_odex_install.mk: update for manifest_check.py changes.
Some of the uses-library entries in the manifest are propagated
transitively from dependencies (which are not necessarily uses-libraries
themselves, they may be e.g. static libraries). In Soong this is done at
deps mutator phase, but in makefiles it is done via dexpreopt.config
files (see build/make/core/dex_preopt_config_merger.py).

This patch updates makefiles for the changes in manifest_check.py:
  - option for dependency dexpreopt.configs is renamed
  - option for self dexpreopt.config is added
  - 'Optional' field is added in makefile-generated dexpreopt.config
  - definitions of local variables are moved before uses-library check,
    as now they are needed there

Bug: 214255490
Test: lunch aosp_cf_x86_64_phone-userdebug && m && launch_cvd \
    && adb wait-for-device && adb root \
    && adb logcat | grep -E 'ClassLoaderContext [a-z ]+ mismatch'
    # empty output, no errors at boot
Change-Id: I6522319a8415f22f90fc96059a34675fb830e5cc